Updated all packages
Simplified patient creation

Simplified therapy creation

General bugfixing and refactoring
