+++ package.json
@@ -8,6 +8,10 @@
     "build": "vite build",
     "preview": "vite preview"
   },
+  "engines": {
+    "node": "22.x",
+    "npm": "10.x"
+  },
   "dependencies": {
     "@vicons/ionicons5": "^0.12.0",
     "axios": "^1.6.7",
