+++ src/components/patient-details/Therapy.vue
@@ -8,7 +8,7 @@ import {
 } from "@vicons/ionicons5";
 import { Therapy } from "@/classes/therapy-dto";
 import { useI18n } from "vue-i18n";
-import TherapyForm from "@/components/patient-details/TherapyForm.vue";
+import TherapyForm from "@/components/PatientDetailDoctor/TherapyForm.vue";
 import * as api from "@/api";
 import { useDialog } from "naive-ui";
 import { useThemeVars } from "naive-ui";
