--- src/components/patient-list/NewPatientForm.vue
@@ -1,198 +1,0 @@
-<script lang="ts">
-import { computed, defineComponent, ref } from "vue";
-import { useI18n } from "vue-i18n";
-import { Close as CloseIcon } from "@vicons/ionicons5";
-import { Save as SaveIcon } from "@vicons/ionicons5";
-import * as api from "@/api";
-
-export default defineComponent({
-  name: "NewPatientForm",
-  components: { CloseIcon, SaveIcon },
-  emits: ["saved"],
-  setup(_, context) {
-    const { t } = useI18n({ useScope: "global", inheritLocale: true });
-    const showSpin = ref(false);
-    const formRef = ref(null);
-    const model = ref({
-      username: undefined as string | undefined,
-      name: undefined as string | undefined,
-      surname: undefined as string | undefined,
-      cf: undefined as string | undefined,
-      email: undefined as string | undefined,
-      phone: undefined as string | undefined,
-      sex: undefined as string | undefined,
-      dob: undefined as Date | undefined,
-      residence: undefined as string | undefined,
-    });
-    const revertChanges = () => {
-      model.value = {
-        username: undefined,
-        name: undefined,
-        surname: undefined,
-        cf: undefined,
-        email: undefined,
-        phone: undefined,
-        sex: undefined,
-        dob: undefined,
-        residence: undefined,
-      };
-    };
-
-    const disableSave = computed(() => {
-      return !model.value.username;
-    });
-
-    const save = () => {
-      showSpin.value = true;
-      console.log("Creating patient");
-      api.administration
-          .createPatient({
-            username: model.value.username as string,
-            user: {
-              name: model.value.name,
-              surname: model.value.surname,
-              cf: model.value.cf,
-              email: model.value.email,
-              phone: model.value.phone,
-              dob: model.value.dob as Date,
-              sex: model.value.sex,
-              residence: model.value.residence,
-            },
-            patient: {}
-          })
-          .then(() => {
-            showSpin.value = false;
-            console.log("Patient created")
-            context.emit("saved");
-          })
-          .catch((error) => {
-            showSpin.value = false;
-            console.error(error);
-          });
-    };
-
-    const rules = computed(() => {
-      return {
-        username: [
-          {
-            required: true,
-            message: t("login.usernameValidation"),
-            trigger: ["input", "blur"],
-          },
-        ],
-      };
-    });
-    return {
-      t,
-      save,
-      disableSave,
-      model,
-      rules,
-      formRef,
-      showSpin,
-      revertChanges,
-    };
-  },
-});
-</script>
-
-
-<template>
-  <n-spin :show="showSpin">
-    <n-form label-placement="left" :model="model" :rules="rules" ref="formRef">
-      <n-grid cols="1 s:1 m:2 l:2 xl:2 2xl:2" :x-gap="24" responsive="screen">
-        <n-form-item-gi :label="t('patients.username')" path="username">
-          <n-input
-            :placeholder="t('patients.username')"
-            v-model:value="model.username"
-          />
-        </n-form-item-gi>
-
-        <n-form-item-gi :label="t('patients.name')">
-          <n-input
-            :placeholder="t('patients.name')"
-            path="name"
-            v-model:value="model.name"
-          />
-
-          <n-input
-            :placeholder="t('patients.surname')"
-            v-model:value="model.surname"
-            path="surname"
-            style="margin-left: 5px"
-          />
-        </n-form-item-gi>
-
-        <n-form-item-gi :label="t('patients.cf')" path="cf">
-          <n-input :placeholder="t('patients.cf')" v-model:value="model.cf" />
-        </n-form-item-gi>
-
-        <n-form-item-gi :label="t('profile.email')" path="email">
-          <n-input
-            :placeholder="t('profile.email')"
-            v-model:value="model.email"
-          />
-        </n-form-item-gi>
-
-        <n-form-item-gi :label="t('profile.phone')" path="phone">
-          <n-input
-            :placeholder="t('profile.phone')"
-            v-model:value="model.phone"
-          />
-        </n-form-item-gi>
-
-        <n-form-item-gi :label="t('profile.residence')" path="residence">
-          <n-input
-            :placeholder="t('profile.residence')"
-            v-model:value="model.residence"
-          />
-        </n-form-item-gi>
-
-        <n-form-item-gi :label="t('profile.userType')">
-          <n-input disabled :placeholder="t('general.patient', 1)" />
-        </n-form-item-gi>
-
-        <n-form-item-gi :label="t('patients.sex')" path="sex">
-          <n-select
-            v-model:value="model.sex"
-            placeholder="Select"
-            :options="[
-              { label: t('patients.male'), value: 'Male' },
-              { label: t('patients.female'), value: 'Female' },
-              { label: t('patients.other'), value: 'Other' },
-            ]"
-          />
-        </n-form-item-gi>
-
-        <n-form-item-gi :label="t('patients.dob')" path="dob">
-          <n-date-picker
-            v-model:value="model.dob"
-            :placeholder="t('patients.dob')"
-            type="date"
-            format="dd-MM-yyyy"
-          />
-        </n-form-item-gi>
-      </n-grid>
-    </n-form>
-
-    <n-space justify="end">
-      <n-button type="error" ghost @click="revertChanges">
-        <template #icon>
-          <n-icon>
-            <close-icon />
-          </n-icon>
-        </template>
-        {{ t("general.undo") }}
-      </n-button>
-
-      <n-button type="primary" @click="save" :disabled="disableSave">
-        <template #icon>
-          <n-icon>
-            <save-icon />
-          </n-icon>
-        </template>
-        {{ t("general.save") }}
-      </n-button>
-    </n-space>
-  </n-spin>
-</template>
