+++ src/languages/locales/it-IT.json
@@ -306,9 +306,9 @@
         "SU":"Domenica|D"
     },
     "therapyForm":{
-        "insertDrug":"1. Inserisci il nome del farmaco",
-        "insertPosology": "2. Inserisci la posologia (1 cp, 20 mg, ...)",
-        "insertDateInterval": "3. Seleziona la data di inizio e di fine dei promemoria per la terapia",
-        "insertDoses": "4. Inserisci gli orari di somministrazione"
+        "insertDrug":"2. Inserisci il nome del farmaco",
+        "insertPosology": "3. Inserisci la posologia (1 cp, 20 mg, ...)",
+        "insertDateInterval": "4. Seleziona la data di inizio e di fine dei promemoria per la terapia",
+        "insertDoses": "1. Inserisci gli orari di somministrazione"
     }
 }
